Fixed infinite loop cycles when running dotnet watch. (#3446)

+++ src/StrawberryShake/SourceGenerator/src/CodeGeneration.CSharp.Analyzers/SingleFileDocumentWriter.cs
@@ -3,15 +3,18 @@ using System.IO;
 using System.Text;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
+using HotChocolate.Language;
 using IOPath = System.IO.Path;
 
 namespace StrawberryShake.CodeGeneration.CSharp.Analyzers
 {
     public class SingleFileDocumentWriter : IDocumentWriter
     {
-        private StringBuilder _content = new();
+        private readonly StringBuilder _content = new();
+        private readonly MD5DocumentHashProvider _hashProvider = new();
         private GeneratorExecutionContext? _execution;
         private string? _fileName;
+        private string? _hashFile;
         private bool _emitCode;
 
         public void WriteDocument(ClientGeneratorContext context, SourceDocument document)
@@ -28,6 +31,9 @@ namespace StrawberryShake.CodeGeneration.CSharp.Analyzers
                 _fileName = IOPath.Combine(
                     context.OutputDirectory,
                     $"{context.Settings.Name}.StrawberryShake.cs");
+                _hashFile = IOPath.Combine(
+                    context.GetStateDirectory(),
+                    context.Settings.Name + ".md5");
                 _emitCode = context.Settings.EmitGeneratedCode;
                 context.FileNames.Add(_fileName);
             }
@@ -51,19 +57,34 @@ namespace StrawberryShake.CodeGeneration.CSharp.Analyzers
 
             if (_emitCode && _fileName is not null)
             {
-                string directory = IOPath.GetDirectoryName(_fileName);
+                string? hash = _hashFile is not null && File.Exists(_hashFile)
+                    ? File.ReadAllText(_hashFile)
+                    : null;
 
-                if (!Directory.Exists(directory))
+                string currentHash = _hashProvider.ComputeHash(
+                    Encoding.UTF8.GetBytes(_content.ToString()));
+
+                bool fileExists = File.Exists(_fileName);
+
+                // we only write the file if it has changed so we do not trigger a loop on
+                // dotnet watch.
+                if (!fileExists || !currentHash.Equals(hash, StringComparison.Ordinal))
                 {
-                    Directory.CreateDirectory(directory);
-                }
+                    string directory = IOPath.GetDirectoryName(_fileName);
 
-                if (File.Exists(_fileName))
-                {
-                    File.Delete(_fileName);
-                }
+                    if (!Directory.Exists(directory))
+                    {
+                        Directory.CreateDirectory(directory);
+                    }
 
-                File.WriteAllText(_fileName, _content.ToString(), Encoding.UTF8);
+                    if (fileExists)
+                    {
+                        File.Delete(_fileName);
+                    }
+
+                    File.WriteAllText(_fileName, _content.ToString(), Encoding.UTF8);
+                    File.WriteAllText(_hashFile, currentHash, Encoding.UTF8);
+                }
             }
         }
     }
